+++ examples/Test.cs
@@ -4,9 +4,9 @@
 var builder = clapnet.CommandBuilder.New();
 return builder
     .UseCamelCase()
-    .With(() => Console.WriteLine("ssss"), "", "lambda_two")
+    .With(() => Console.WriteLine("ssss"), "Other function to call", "lambda_two")
     .With(Gather, "Documentation for gather command")
-    .With(Failing, "This command will fail")
+    .With(Failing)
     .With(() => Console.WriteLine("ssss"), "Test command", "lambda")
     .WithRootCommand(Other, "Super command to show what can be done")
     .Run(args);
@@ -22,20 +22,26 @@ void Other(SomeSettings settings)
     Console.WriteLine("Hello World from the root command");
 }
 
+/// <summary>
+/// This command will fail
+/// </summary>
 int Failing()
 {
     /// Returning non zero value
     return 1;
 }
 
+/// <summary>
+/// Config
+/// </summary>
 class SomeSettings
 {
     /// <summary>
-    /// Test value, it will be possible to set its value by passing `--test true/false`.
+    /// Test value.
     /// </summary>
-    public bool Test = false;
+    public bool TestV = false;
     /// <summary>
-    /// Test value, it will be possible to set its value by passing `--other "string value"`.
+    /// Test value.
     /// </summary>
     public string other = "Default Value";